refactor: migrate styles to tailwindcss
