+++ package.json
@@ -27,6 +27,7 @@
     "@syfxlin/reks": "^1.0.4",
     "@tippyjs/react": "^4.2.6",
     "artalk": "^2.9.1",
+    "clsx": "2.1.1",
     "fs-extra": "^11.3.5",
     "fuse.js": "^7.3.0",
     "gh-lang-colors": "^0.1.21",
@@ -47,6 +48,7 @@
     "sharp": "^0.34.5",
     "shikiji": "0.10.2",
     "swr": "^2.4.1",
+    "tailwind-merge": "^3.6.0",
     "tippy.js": "^6.3.7"
   },
   "devDependencies": {
