+++ src/components/docs/message/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { ReactNode } from "react";
-import { cx } from "../../../utils/class-name";
+import { cx } from "../../../utils/styles";
 
 export interface MessageProps {
   type: "warn" | "info" | "error" | "success";
