+++ src/components/layouts/hero/index.tsx
@@ -2,7 +2,7 @@ import { Caveat } from "next/font/google";
 import * as React from "react";
 import { fetcher } from "../../../contents";
 import { t } from "../../../locales";
-import { cx } from "../../../utils/class-name";
+import { cx } from "../../../utils/styles";
 import { Image } from "../../ui/image";
 
 const caveat = Caveat({ subsets: ["latin"] });
