+++ src/components/layouts/main/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { PropsWithChildren } from "react";
-import { cx } from "../../../utils/class-name";
+import { cx } from "../../../utils/styles";
 
 export type MainProps = PropsWithChildren<{
   size?: "sm" | "md" | "lg";
